Move CategorySelectorItem to component
